+++ examples/resources/foundry_vector_store/resource.tf
@@ -1,0 +1,7 @@
+resource "foundry_vector_store" "support" {
+  name = "support-knowledge-base"
+  metadata = {
+    team = "support"
+  }
+  expires_after_days = 30
+}